♻️ 関数分けした
